Crate sp_evm_tracker

Inherents for EVM tracker

Structs

BlockGasLimit

EVM block gas limit is set to maximum to allow all the transaction stored on Consensus chain.

GasLimitPovSizeRatio

GasPerByte

Gas per byte Ethereum’s Yellow Paper states that it costs 20,000 gas to store one 256-bit word. 1 Byte costs 20_000/32 = 625

InherentDataProvider

Provides the set code inherent data.

InherentType

The type of the Subspace inherent data.

StorageFeeRatio

Proportion of final (gas_price * gas_used) given as storage fee.

WeightPerGas

Enums

InherentError

Constants

GAS_PER_SECOND

Current approximation of the gas/s consumption considering EVM execution over compiled WASM (on 4.4Ghz CPU).

INHERENT_IDENTIFIER

Executive inherent identifier.

WEIGHT_PER_GAS

Approximate ratio of the amount of Weight per Gas. u64 works for approximations because Weight is a very small unit compared to gas.

Traits

EvmTrackerApi

Api to check and verify the evm-tracker extrinsic calls